Fix bug with includes
We used to load .bashrc in order to add the path to `stddef.h' to C_INCLUDE_PATH and CPLUS_INCLUDE_PATH. OptiTrust needs to know where to find this header which is a dependency of `stdlib.h'. However, loading of the .bashrc prevents `omp-apac/concom' from compiling, i.e. the compiler ignores `stdatomic.h'. To make everyone happy, we tell OptiTrust where to find `stddef.h' using the `-I' option in Makefile.